[migration/WF-149]: migrate feature auth
