+++ src/entities/User/index.ts
@@ -1,4 +1,4 @@
-export type { IUserTabs } from './types';
+export type { IUserTabs, UserResponse } from './types';
 export { FormTemplate, ProfileTemplate, ViewPersonalDetail } from './components';
 export { useProfile } from './hooks';
 export { UserService } from './user.service';
