+++ src/entities/User/types/index.ts
@@ -1,1 +1,2 @@
 export type { IUserTabs } from './IUserTabs';
+export type { UserResponse } from './UserResponse';
