+++ src/features/Auth/components/LoginForm/LoginForm.module.scss
@@ -1,14 +1,23 @@
-.errorMessage {
-	color: var(--danger-main);
-}
-
 .form {
 	margin-top: 46px;
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	justify-content: center;
 
-	@apply flex-col-center-center;
+	.control {
+		width: 100%;
+		max-width: 480px;
+	}
 
 	.input {
+		width: 100%;
 		max-width: 480px;
-		@apply mt-16 w-full;
+		margin-top: 52px;
+	}
+
+	.button {
+		width: 100%;
+		margin-top: 40px;
 	}
 }
